fix: remove --dev from install command (#20)
TailwindCSS currently recommends installation as a standard dependency
per [Next.js install
guide](https://tailwindcss.com/docs/installation/framework-guides/nextjs).
This commit updates the install command to match TailwindCSS for
consistency.

If the current dependency type is retained, --save-dev or -D is
recommended per [NPM
docs](https://docs.npmjs.com/specifying-dependencies-and-devdependencies-in-a-package-json-file#adding-dependencies-to-a-packagejson-file-from-the-command-line).